pyproject.toml
[tool.poetry]
name = "gridder"
version = "0.1.0"
description = "A script to divide an image by a specified grid size."
authors = "Carlos Rayon"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"